Move Polygon distance code into Distance trait
